+++ src/Client/Update/UpdateUtil.fs
@@ -38,10 +38,76 @@ module JsonHelper =
             datamapParent, datamap
         )
 
+module JsonGenericHelper =
+
+    let toFileName (id: string) (fileType: ArcFilesDiscriminate) (jsonType: JsonExportFormat) =
+        let n = System.DateTime.Now.ToUniversalTime().ToString("yyyyMMdd_hhmmss")
+        let formatString = jsonType.ToString()
+        let fileTypeString = fileType.ToString()
+        n + "_" + fileTypeString + "_" + id + "_" + formatString + ".json"
+
+    let tryParseJsonFileName (fileName: string) =
+        if fileName.EndsWith(".json") then
+            let parts = fileName.Substring(0, fileName.Length - 5).Split("_")
+
+            let jsonFormat =
+                parts |> Array.tryPick (fun part -> JsonExportFormat.tryFromString part)
+
+            let arcfile =
+                parts |> Array.tryPick (fun part -> ArcFilesDiscriminate.tryFromString part)
+
+            match jsonFormat, arcfile with
+            | Some jf, Some af -> Some(jf, af)
+            | _ -> None
+        else
+            None
+
 module JsonImportHelper =
 
     open ARCtrl
     open FileImport
+
+    let tryParseFromJsonString
+        (
+            jsonString: string,
+            jsonType: JsonExportFormat option,
+            filetype: ArcFilesDiscriminate option,
+            fileName: string option
+        ) : ArcFiles option =
+        let assumedJsonType =
+            match jsonType, filetype with
+            | Some jt, Some ft -> (jt, ft) |> Some
+            | _, _ ->
+                match fileName with
+                | Some name ->
+                    let resOpt = JsonGenericHelper.tryParseJsonFileName name
+
+                    match resOpt with
+                    | Some(jf, af) -> Some(jf, af)
+                    | None -> None
+                | None -> None
+
+        match assumedJsonType with
+        | Some(jsonFormat, arcfileType) ->
+            let arcfile =
+                Spreadsheet.IO.Json.readFromJsonMap.[(arcfileType, jsonFormat)] jsonString
+
+            Some arcfile
+        | None -> None
+
+    let parseFromJsonString
+        (
+            jsonString: string,
+            jsonType: JsonExportFormat option,
+            filetype: ArcFilesDiscriminate option,
+            fileName: string option
+        ) : ArcFiles =
+        match tryParseFromJsonString (jsonString, jsonType, filetype, fileName) with
+        | Some arcfile -> arcfile
+        | None ->
+            failwith
+                "Error. Unable to find correct JSON format. This function relies on correct naming conventions for the file. We will improve this in the future. The file name must contain the json format, as well as the file type, separated by \"_\". Example: 'ARCtrlCompressed_Assay.json"
+
 
     /// <summary>
     ///
@@ -209,8 +275,8 @@ module JsonExportHelper =
     /// <param name="jef"></param>
     let parseToJsonString (arcfile: ArcFiles, jef: JsonExportFormat) =
         let name, jsonString =
-            let n = System.DateTime.Now.ToUniversalTime().ToString("yyyyMMdd_hhmmss")
-            let nameFromId (id: string) = (n + "_" + id + ".json")
+            let nameFromId (id: string) =
+                JsonGenericHelper.toFileName id arcfile.RelatedArcFilesDiscriminate jef
 
             match arcfile, jef with
             | Investigation ai, JsonExportFormat.ARCtrl -> nameFromId ai.Identifier, ArcInvestigation.toJsonString 0 ai
@@ -250,7 +316,7 @@ module JsonExportHelper =
             | Workflow _, anyElse ->
                 failwithf "Error. It is not intended to parse Workflow to %s format." (string anyElse)
 
-            | DataMap(_, d), JsonExportFormat.ARCtrl -> n + "_datamap", DataMap.toJsonString 0 d
+            | DataMap(_, d), JsonExportFormat.ARCtrl -> nameFromId "", DataMap.toJsonString 0 d
             | DataMap(_), anyElse ->
                 failwithf "Error. It is not intended to parse Datamap to %s format." (string anyElse)
         // | _ -> failwith $"Error, the selected type {arcfile} is not supported to be exported." //Have to implement the logic for run when toJsonString has been implemented for it